requirements.txt
pandas
python-dotenv
aiofiles
loguru
tqdm
pillow
tesserocr
opencv-python-headless